Move e2e tests to provider
