+++ packages/api/test/e2e/subscribe.test.js
@@ -1,0 +1,37 @@
+// Copyright 2017-2018 @polkadot/api-provider authors & contributors
+// This software may be modified and distributed under the terms
+// of the ISC license. See the LICENSE file for details.
+
+import storage from '@polkadot/storage/testing';
+
+import Ws from '../../../api-provider/src/ws';
+import Api from '../../src';
+
+describe.skip('e2e subscriptionsw', () => {
+  let api;
+
+  beforeEach(() => {
+    jest.setTimeout(30000);
+    api = new Api(new Ws('ws://127.0.0.1:9944'));
+  });
+
+  it('retrieves current timestamp', (done) => {
+    let count = 0;
+
+    return api.state
+      .storage([[storage.timestamp.now]], (error, data) => {
+        if (error) {
+          return done(error);
+        }
+
+        expect(data).toBeDefined();
+
+        if (++count === 3) {
+          done();
+        }
+      })
+      .then((subscriptionId) => {
+        console.log('newHead: subscriptionId =', subscriptionId);
+      });
+  });
+});
